Implemented the dynamic fib generator.

+++ app/src/Main.cpp
@@ -1,16 +1,18 @@
 #include <vector>
 #include <stdexcept>
+#include <iostream>
+#include <stdint.h>
 
-std::vector<int> fibVector{0, 1};
+std::vector<u_int64_t> fibVector{0, 1};
 
 //Increases capacity of vector by a pre-configured step size
-void steppedIncreaseOfCapacityOfVector(std::vector<int> *ptr) {
+void steppedIncreaseOfCapacityOfVector(std::vector<u_int64_t> *ptr) {
     const int CAPACITYINCREASESTEPSIZE = 100;
     int newVectorCapacity= ptr->capacity() + CAPACITYINCREASESTEPSIZE;
     ptr->reserve(newVectorCapacity);
 } 
 
-int getNthFibNumber(int n) {
+u_int64_t getNthFibNumber(u_int64_t n) {
     if(n < 0) {
         throw std::invalid_argument("n must be 0 or greater.");
     }
@@ -22,7 +24,7 @@ int getNthFibNumber(int n) {
         else {
             //n-2 needs to be called first, since otherwise the vector will
             //be out of order after the val is push_back onto vector
-            int thisFibNumber = getNthFibNumber(n - 2) + getNthFibNumber(n - 1);
+            u_int64_t thisFibNumber = getNthFibNumber(n - 2) + getNthFibNumber(n - 1);
             
             //Increase the capacity by one step if needed
             if(n + 1 > fibVector.capacity()) {
